[#130] SortOption 분리
